refactor: Moving Context to new file
